thingpedia/api: remove login-dependent APIs from /thingpedia/api
/thingpedia/api is the public API, authenticated through OAuth
or developer keys. Anything cookie/session-authenticated should
be elsewhere, where we can apply the correct security protections.
